+++ scoped-gc/CHANGELOG.md
@@ -1,3 +1,7 @@
+## Next
+
+- **[Feature]** Implement `Trace` for `::std::ops::Range`.
+
 ## 0.1.4 (2019-07-30)
 
 - **[Fix]** **Fix use-after-free** caused by invalid lifetime constraint.